Plan renovation of wasi-stage2
The interpreter in this stage is okay but it is very much mixed with its
real payload. We want to pass control to _some_ WASI module as quick as
possible which will be stage3. But there's no need for this to be the
final stage.

Indeed, it would also interesting to bootstrap further in a Rust/WASI
implementation of configuration file and unzip.